Set Custom Labels

This function allows for setting global options for how rx should label individual terms in a formula. The default labels are lhs for the response or dependent variables, and the rhs are the terms or independent variables, which require no shorthand or wrappers to work. This will exist for the reminder of the working session.

set_rx_labels(labels)

set_rx_theme(theme)

reset_rx_labels()

Arguments

labels
A named list pair of character vectors that should be used to "wrap" formula terms, and change their labels. The default options labels the independent variables as rhs and the dependent variables as lhs.

theme
A preset list of labels customized for specific packages. These can be expanded via pull request.

Currently supported themes:

package	lhs	rhs	labels

murmur	outcomes	predictors	X = exposure; F = fixed

ggdag	lhs	rhs	X = exposure; Y = "outcome", L = "latent"

Value

Changes to global options within the working environment.

Details

To modify the defaults, simply provide a named list pair in the argument labels that renames lhs and rhs. For example, the default labels can be revised using list(lhs = "outcomes", rhs = "predictors").

To add new labels, add additional named/paired arguments. The name serves as the wrapper for the formula, creating a shorthand for adding labels. For example, using list(X = "exposure") will label any component of a formula as an "exposure" variable. To use the wrapping feature, simply use the name of the list argument in parentheses around the individual term, as below.

$$mpg ~ X(wt) + hp + cyl$$